fix(quic): add custom cert verifier to disable SNI for mTLS
The TLS client uses a custom verifier that disables SNI since libp2p nodes use dynamic IPs and
certificates are based on peer identity rather than hostnames. The TLS `upgrade` already solved this
for the TCP transport but QUIC doesn't use that, which is why we need a custom verifier to disable
SNI while maintain the rest.